Add army units

+++ Assets/Scripts/Core/Army/UnitInfo.cs
@@ -18,7 +18,7 @@ namespace EuropeanWars.Core.Army {
         public int health;
         public int attackCooldown;
         public int speed;
-        public int maintenance;
+        public float maintenance;
 
         public UnitInfo(UnitData data) {
             this.data = data;
